workspaces/ecommerce_clean/requirements.txt
PyJWT>=2.8.0
asyncio
prometheus-client>=0.19.0
pydantic>=2.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0